graphql-alt: include page in tx_bounds (#23488)
## Description 

1. Move `tx_bounds` into its own file instead of being part of the
`transaction` functionality.
2. Refactor page bounds calculation into `tx_bounds` instead of having
mostly duplicate logic for transactions and events.

## Test plan 

Functionality is the same so existing tests cover changes.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
